+++ src/partners/dexAg/config/config.js
@@ -21,6 +21,7 @@ const WETH_TOKEN_ADDRESS = '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2';
 
 const SUPPORTED_DEXES = [
   'uniswap',
+  'uniswap_v2',
   'bancor',
   'kyber',
   'zero_x',